- A bunch of various fixes and consistency junk  - Some were reported by Diamondo25, some were fixed by the same
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/trunk@2890 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/ChannelServer/Fame.cpp
@@ -24,7 +24,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 void Fame::handleFame(Player *player, PacketReader &packet) {
 	int32_t playerid = packet.get<int32_t>();
-	uint8_t type = (uint8_t) packet.get<int32_t>();
+	uint8_t type = packet.get<uint8_t>();
 	if (player->getId() > 0) {
 		if (player->getId() != playerid) {
 			int32_t checkResult = canFame(player, playerid);
